feat: consume design system in remix project
