+++ examples/remix/package.json
@@ -21,6 +21,7 @@
     "@remix-run/react": "^1.17.1",
     "@remix-run/serve": "^1.17.1",
     "@tokenami/css": "workspace:*",
+    "@tokenami/example-design-system": "workspace:*",
     "isbot": "^3.6.8",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
