+++ packages/config/src/index.ts
@@ -1,4 +1,4 @@
-export { type DefaultConfig, default as defaultConfig } from './config.default';
+export * from './config.default';
 export * from './config';
 export * from './supports';
 export * from './shorthands';
